# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for extracting study documents and generating interactive quizzes.
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  code?: string;
}

export interface DocumentUpload {
  file: Blob;
}

export interface ExtractedDocument {
  fileName: string;
  fileType: string;
  characterCount: number;
  wordCount: number;
  chunkCount: number;
  text: string;
  /** @nullable */
  warning: string | null;
}

export interface QuizGenerationInput {
  /**
     * @minLength 80
     * @maxLength 120000
     */
  text: string;
  fileName?: string;
  /**
     * @minimum 3
     * @maximum 10
     */
  questionCount: number;
}

export interface QuizQuestion {
  id: string;
  question: string;
  /**
     * @minItems 4
     * @maxItems 4
     */
  options: string[];
  /**
     * @minimum 0
     * @maximum 3
     */
  correctAnswerIndex: number;
  explanation: string;
}

export interface Quiz {
  title: string;
  sourceFileName: string;
  questions: QuizQuestion[];
  generatedWith: string;
  isDemo: boolean;
}