# lib/api-zod/src/generated/types/errorResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for extracting study documents and generating interactive quizzes.
 * OpenAPI spec version: 0.1.0
 */

export interface ErrorResponse {
  error: string;
  code?: string;
}