# lib/api-zod/src/generated/types/extractedDocument.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for extracting study documents and generating interactive quizzes.
 * OpenAPI spec version: 0.1.0
 */

export interface ExtractedDocument {
  fileName: string;
  fileType: string;
  characterCount: number;
  wordCount: number;
  chunkCount: number;
  text: string;
  /** @nullable */
  warning: string | null;
}