# lib/api-zod/src/generated/types/quiz.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for extracting study documents and generating interactive quizzes.
 * OpenAPI spec version: 0.1.0
 */
import type { QuizQuestion } from './quizQuestion';

export interface Quiz {
  title: string;
  sourceFileName: string;
  questions: QuizQuestion[];
  generatedWith: string;
  isDemo: boolean;
}