# lib/api-zod/src/generated/types/quizGenerationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for extracting study documents and generating interactive quizzes.
 * OpenAPI spec version: 0.1.0
 */

export interface QuizGenerationInput {
  /**
     * @minLength 80
     * @maxLength 120000
     */
  text: string;
  fileName?: string;
  /**
     * @minimum 3
     * @maximum 10
     */
  questionCount: number;
}