# lib/api-zod/src/generated/types/quizQuestion.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for extracting study documents and generating interactive quizzes.
 * OpenAPI spec version: 0.1.0
 */

export interface QuizQuestion {
  id: string;
  question: string;
  /**
     * @minItems 4
     * @maxItems 4
     */
  options: string[];
  /**
     * @minimum 0
     * @maximum 3
     */
  correctAnswerIndex: number;
  explanation: string;
}